compatible PHP8 of wechatpay-apiv3/wechatpay-guzzle-middleware#51
